Cleanup opencl_backend.cpp  ready for syncing buffers.
